requirements.txt
requests==2.26.0
python-vlc==3.0.11115
moviepy==1.0.3
Flask==2.1.1
live555==1.0.0
